+++ src/main/kotlin/com/odtheking/odin/clickgui/settings/impl/HUDSetting.kt
@@ -27,9 +27,9 @@ class HUDSetting(
 
     constructor(
         name: String,
-        x: Float,
-        y: Float,
-        scale: Float,
+        x: Int,
+        y: Int,
+        scale: Int,
         toggleable: Boolean,
         description: String,
         module: Module,
@@ -95,9 +95,9 @@ class HUDSetting(
 
     override fun read(element: JsonElement?) {
         if (element !is JsonObject) return
-        value.x = element.get("x")?.asFloat ?: value.x
-        value.y = element.get("y")?.asFloat ?: value.y
-        value.scale = element.get("scale")?.asFloat ?: value.scale
+        value.x = element.get("x")?.asInt ?: value.x
+        value.y = element.get("y")?.asInt ?: value.y
+        value.scale = element.get("scale")?.asInt ?: value.scale
         value.enabled = if (toggleable) element.get("enabled")?.asBoolean ?: value.enabled else true
     }
 }
